update to 1.12.1

+++ src/main/java/com/bymarcin/openglasses/surface/widgets/core/attribute/ILookable.java
@@ -1,10 +1,10 @@
 package com.bymarcin.openglasses.surface.widgets.core.attribute;
 
 public interface ILookable extends IAttribute{
-	public void setLookingAt(int x, int y, int z);
+	public void setLookingAt(double x, double y, double z);
 	public boolean isLookingAtEnable();
 	public void setLookingAtEnable(boolean enable);
-	public int getLookingAtX();
-	public int getLookingAtY();
-	public int getLookingAtZ();
+	public double getLookingAtX();
+	public double getLookingAtY();
+	public double getLookingAtZ();
 }
